+++ src/activity.h
@@ -618,8 +618,8 @@ public:
 
   double run(Arrival* arrival) {
     double value = get<double>(delay, 0, arrival);
-    if (R_IsNA(value))
-      Rcpp::stop("%s: missing value (NA returned)", name);
+    if (ISNAN(value))
+      Rcpp::stop("%s: missing value (NA or NaN returned)", name);
     return std::abs(value);
   }