+++ tests/testthat/test-trajectory-timeout.R
@@ -13,10 +13,14 @@ test_that("a task function that returns a non-numeric value fails", {
 test_that("a missing value fails", {
   t0 <- trajectory() %>%
     timeout(NA)
-
   env <- simmer(verbose = TRUE) %>%
     add_generator("dummy", t0, at(0))
+  expect_error(env %>% run())
 
+  t0 <- trajectory() %>%
+    timeout(NaN)
+  env <- simmer(verbose = TRUE) %>%
+    add_generator("dummy", t0, at(0))
   expect_error(env %>% run())
 })