+++ NEWS.md
@@ -7,7 +7,7 @@
 * Apply `select()`'s `policy` also when `resources` is a function (#92).
 * Accept dynamic timeouts in batches (#93).
 * Change `rollback()`'s default behaviour to `times=Inf`, i.e., infinite loop (#95).
-* Stop and throw an error when `timeout()` returns a missing value (#96).
+* Stop and throw an error when `timeout()` returns a missing value (#96 and #97).
 
 # simmer 3.6.1